change status endpoint to receipt
