+++ packages/webapi/model/reqstatus.go
@@ -6,7 +6,7 @@ type WaitRequestProcessedParams struct {
 	Timeout time.Duration `swagger:"desc(Timeout in nanoseconds),default(30 seconds)"`
 }
 
-type RequestStatusResponse struct {
+type RequestReceiptResponse struct {
 	IsProcessed bool `swagger:"desc(True if the request has been processed)"`
 }
